Fix Decimal to float conversion in price drop evaluator

+++ meridian/signals/evaluators/price_drop.py
@@ -18,7 +18,6 @@ class PriceDrop30dEvaluator(SignalEvaluator):
         history: Sequence[MarketReport],
     ) -> SignalResult:
         """Evaluate price drop over 30 days."""
-        # Get reports from last 30 days
         cutoff = date.today() - timedelta(days=30)
         recent_reports = [
             r for r in history
@@ -34,21 +33,17 @@ class PriceDrop30dEvaluator(SignalEvaluator):
                 fired=False,
             )
 
-        # Sort by date
         recent_reports.sort(key=lambda r: r.report_date)
 
-        # Calculate price change
         oldest_price = recent_reports[0].median_price
         newest_price = recent_reports[-1].median_price
 
         if oldest_price == 0:
             price_change_pct = 0.0
         else:
-            price_change_pct = ((newest_price - oldest_price) / oldest_price) * 100
+            price_change_pct = float(((newest_price - oldest_price) / oldest_price) * 100)
 
-        # Threshold from config (default 5%)
-        threshold = -5.0  # TODO: load from config
-
+        threshold = -5.0
         fired = price_change_pct <= threshold
 
         return SignalResult(
@@ -59,6 +54,6 @@ class PriceDrop30dEvaluator(SignalEvaluator):
                 "threshold": threshold,
                 "days_window": 30,
             },
-            confidence=min(1.0, len(recent_reports) / 10.0),  # Higher confidence with more data
+            confidence=min(1.0, len(recent_reports) / 10.0),
             fired=fired,
         )
